% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_libdatabases.R
\name{get.libdatabases}
\alias{get.libdatabases}
\title{Get the databases of the Retip library}
\usage{
get.libdatabases(include_metadata = FALSE)
}
\arguments{
\item{include_metadata}{The output includes the molecules information. It defaults to FALSE.}
}
\value{
The databases of the Retip library
}
\description{
Extract the databases included in the library
}
\examples{
get.libdatabases(include_metadata = FALSE)
}
