% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_libdescriptors.R
\name{get.libdescriptors}
\alias{get.libdescriptors}
\title{Get the descriptors of the Retip library}
\usage{
get.libdescriptors(include_metadata = FALSE)
}
\arguments{
\item{include_metadata}{The output includes the molecules information. It defaults to FALSE.}
}
\value{
The descriptors of the Retip library
}
\description{
Extract the descriptors included in the library
}
\examples{
get.libdescriptors(include_metadata = FALSE)
}
