% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_libmetadata.R
\name{get.libmetadata}
\alias{get.libmetadata}
\title{Get the metadata of the Retip library}
\usage{
get.libmetadata()
}
\value{
The metadata of the Retip library
}
\description{
Extract the information (name, InChIKey, mass, formula and smiles) of the molecules included in the library
}
\examples{
get.libmetadata()
}
