% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get.retiplib}
\alias{get.retiplib}
\title{Build the retip library}
\usage{
get.retiplib()
}
\value{
the retip library by merging the three main files
}
\description{
Build the retip library
}
\examples{
fet.retiplib()
}
\references{
xxx
}
\keyword{16}
\keyword{Chemical}
\keyword{Retip}
\keyword{calculated}
\keyword{database}
\keyword{descriptos}
\keyword{different}
\keyword{more}
\keyword{than}
\keyword{with}
