% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retip_lib}
\alias{retip_lib}
\title{Build the Retip library}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3 rows and 1 columns.
}
\usage{
data(retip_lib_1)
}
\value{
The Retip library
}
\description{
Merge the three data objects to build the Retip library, which contains molecules from different databases with descriptors calculated.
}
\examples{
get.retiplib()
}
\keyword{datasets}
