% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_3.R
\docType{data}
\name{retip_lib_3}
\alias{retip_lib_3}
\title{Retip library}
\format{
An object of class \code{data.frame} with 183879 rows and 267 columns.
}
\usage{
data(retip_lib_3)
}
\description{
Third third of the Retip library. Needs to be merged to the other two thirds.
}
\examples{
data(retip_lib_3)
}
\references{
xxx
}
\keyword{16}
\keyword{Chemical}
\keyword{Retip}
\keyword{calculated}
\keyword{database}
\keyword{descriptos}
\keyword{different}
\keyword{more}
\keyword{than}
\keyword{with}
