% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retip_lib_v1}
\alias{retip_lib_v1}
\title{Retip library}
\format{An object of class \code{data.frame} with 321342 rows and 311 columns.}
\usage{
data(retip_lib_v1)
}
\description{
Retip library
}
\examples{
data(retip_lib_v1)
}
\references{
xxx
}
\keyword{16}
\keyword{Chemical}
\keyword{Retip}
\keyword{calculated}
\keyword{database}
\keyword{descriptos}
\keyword{different}
\keyword{more}
\keyword{than}
\keyword{with}
