% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_1.R
\docType{data}
\name{retip_lib_v2_1}
\alias{retip_lib_v2_1}
\title{Retip library}
\format{
An object of class \code{data.frame} with 183879 rows and 267 columns.
}
\usage{
retip_lib_v2_1
}
\description{
Retip library
}
\examples{
data(retip_lib_v2_1)
}
\references{
xxx
}
\keyword{16}
\keyword{Chemical}
\keyword{Retip}
\keyword{calculated}
\keyword{database}
\keyword{descriptos}
\keyword{different}
\keyword{more}
\keyword{than}
\keyword{with}
