% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get.retiplib}
\alias{get.retiplib}
\title{Build the Retip library}
\usage{
get.retiplib()
}
\value{
The Retip library
}
\description{
Build the Retip library by merging three data files that contain molecules from different databases with descriptors calculated.
}
\examples{
get.retiplib()
}
